+++ apps/api/src/modules/research-campaign/research-campaign.module.ts
@@ -1,12 +1,13 @@
 import { Module } from '@nestjs/common';
 import { ExperimentsModule } from '../experiments/experiments.module';
+import { CampaignController } from './campaign.controller';
 import { CampaignReportService } from './campaign-report.service';
 import { ResearchCampaignController } from './research-campaign.controller';
 import { ResearchCampaignService } from './research-campaign.service';
 
 @Module({
   imports: [ExperimentsModule],
-  controllers: [ResearchCampaignController],
+  controllers: [ResearchCampaignController, CampaignController],
   providers: [ResearchCampaignService, CampaignReportService],
   exports: [ResearchCampaignService, CampaignReportService],
 })